FIX: Autolocalize already cooked posts.
Summary:
If a user landed directly on a topic when opening the forum
(e.g. using a link from an email), the links were often not
localized because the posts were cooked before this component
was activated (i.e. before we register callback via api.decorateCooked)

The ugly fix is to apply the component manually to already cooked posts
in the initializer.

Test Plan:
  - Open
    https://international-forum.khanacademy.org/t/new-email-parent-dashboard-strings-please-translate-by-march-11/2145
  - Try F5, links are probably not localized
  - Install the updated component via
    https://international-forum.khanacademy.org/admin/customize/themes
  - Hit F5, links should be localized